Moved server tests into tests; Moved server main() into examples/cli-server.rs
