+++ Cargo.toml
@@ -4,25 +4,21 @@ version = "0.1.1"
 authors = ["Aaron Miller <aaron.miller04@gmail.com>", "manghi <manghirs@gmail.com>"]
 edition = "2018"
 
-[[bin]]
-name = "server"
-path = "src/server.rs"
-
-[[bin]]
-name = "cli-client"
-path = "src/client.rs"
+#[[bin]]
+#name = "cli-client"
+#path = "src/cli-client.rs"
 
 [dependencies]
 futures      = "0.1.17"
 tokio-core   = "0.1.9"
-log          = "*"
-chrono       = "*"
-env_logger   = "0.6.0"
-serde        = "*"
+log          = "0.4.6"
+chrono       = "0.4.6"
+env_logger   = "0.6.1"
+serde        = "1.0.89"
 serde_derive = "1.0.15"
 bincode      = "0.9.2"
-base64       = "*"
-rand         = "*"
+base64       = "0.10.1"
+rand         = "0.6.5"
 time         = "0.1"
 proptest     = "0.7.2"
 semver       = "0.9.0"
