+++ src/server.rs
@@ -17,60 +17,38 @@
  * this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-#[macro_use]
-extern crate serde_derive;
-#[macro_use]
-extern crate log;
-extern crate env_logger;
-extern crate futures;
-extern crate tokio_core;
-extern crate base64;
-extern crate rand;
-extern crate semver;
-#[cfg(test)]
-#[macro_use]
-extern crate proptest;
-extern crate chrono;
-
-#[macro_use]
-mod net;
-
 use crate::net::{
-    RequestAction, ResponseCode, Packet, LineCodec,
+    RequestAction, ResponseCode, Packet,
     UniUpdateType, BroadcastChatMessage, NetworkManager,
-    NetworkQueue,
+    NetworkQueue, get_version, VERSION, has_connection_timed_out
 };
 
 use std::error::Error;
-use std::io::{self, ErrorKind, Write};
-use std::iter;
+use std::io::{self, ErrorKind};
 use std::net::SocketAddr;
-use std::process::exit;
 use std::time::{Duration, Instant};
 use std::collections::HashMap;
 use std::fmt;
 use std::time;
 use std::collections::VecDeque;
-use futures::{Future, Sink, Stream, stream, future::ok, sync::mpsc};
-use tokio_core::reactor::{Core, Timeout};
-use rand::RngCore;
-use semver::Version;
-use chrono::Local;
-use log::LevelFilter;
 
-const TICK_INTERVAL_IN_MS:    u64      = 10;
-const NETWORK_INTERVAL_IN_MS: u64      = 100;    // Arbitrarily chosen
-const HEARTBEAT_INTERVAL_IN_MS: u64    = 1000;    // Arbitrarily chosen
-const MAX_ROOM_NAME:          usize    = 16;
-const MAX_NUM_CHAT_MESSAGES:  usize    = 128;
-const MAX_AGE_CHAT_MESSAGES:  usize    = 60*5; // seconds
-const SERVER_ID:              PlayerID = PlayerID(u64::max_value()); // 0xFFFF....FFFF
+use crate::futures::sync::mpsc;
+use crate::rand::RngCore;
+use crate::semver::Version;
+
+pub const TICK_INTERVAL_IN_MS:    u64      = 10;
+pub const NETWORK_INTERVAL_IN_MS: u64      = 100;    // Arbitrarily chosen
+pub const HEARTBEAT_INTERVAL_IN_MS: u64    = 1000;    // Arbitrarily chosen
+pub const MAX_ROOM_NAME:          usize    = 16;
+pub const MAX_NUM_CHAT_MESSAGES:  usize    = 128;
+pub const MAX_AGE_CHAT_MESSAGES:  usize    = 60*5; // seconds
+pub const SERVER_ID:              PlayerID = PlayerID(u64::max_value()); // 0xFFFF....FFFF
 
 #[derive(PartialEq, Debug, Clone, Copy, Eq, Hash)]
-struct PlayerID(u64);
+pub struct PlayerID(pub u64);
 
 #[derive(PartialEq, Debug, Clone, Copy, Eq, Hash)]
-struct RoomID(u64);
+pub struct RoomID(pub u64);
 
 impl fmt::Display for PlayerID {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
@@ -85,20 +63,20 @@ impl fmt::Display for RoomID {
 }
 
 #[derive(PartialEq, Debug, Clone)]
-struct Player {
-    player_id:     PlayerID,
-    cookie:        String,
-    addr:          SocketAddr,
-    name:          String,
-    request_ack:   Option<u64>,          // The next number we expect is request_ack + 1
-    next_resp_seq: u64,                  // This is the sequence number for the Response packet the Server sends to the Client
-    game_info:     Option<PlayerInGameInfo>,   // none means in lobby
-    heartbeat:     Option<time::Instant>,
+pub struct Player {
+    pub player_id:     PlayerID,
+    pub cookie:        String,
+    pub addr:          SocketAddr,
+    pub name:          String,
+    pub request_ack:   Option<u64>,          // The next number we expect is request_ack + 1
+    pub next_resp_seq: u64,                  // This is the sequence number for the Response packet the Server sends to the Client
+    pub game_info:     Option<PlayerInGameInfo>,   // none means in lobby
+    pub heartbeat:     Option<time::Instant>,
 }
 
 // info for a player as it relates to a game/room
 #[derive(PartialEq, Debug, Clone)]
-struct PlayerInGameInfo {
+pub struct PlayerInGameInfo {
     room_id: RoomID,
     chat_msg_seq_num: Option<u64>,    // Server has confirmed the client has received messages up to this value.
     //XXX PlayerGenState ID within Universe
@@ -106,7 +84,7 @@ struct PlayerInGameInfo {
 }
 
 impl Player {
-    fn increment_response_seq_num(&mut self) -> u64 {
+    pub fn increment_response_seq_num(&mut self) -> u64 {
         let old_seq = self.next_resp_seq;
         self.next_resp_seq += 1;
         old_seq
@@ -115,7 +93,7 @@ impl Player {
     // Update the Server's record of what chat messsage the player has obtained.
     // If the player is in a game, and the player has seen newer chat messages since the last time
     // they updated us on what messages they had, save their sequence number.
-    fn update_chat_seq_num(&mut self, opt_chat_seq_num: Option<u64>) {
+    pub fn update_chat_seq_num(&mut self, opt_chat_seq_num: Option<u64>) {
         if self.game_info.is_none() {
             return;
         }
@@ -130,7 +108,7 @@ impl Player {
     // where N is the last chat message the player has
     // notified the Server it got.
     // Otherwise, None
-    fn get_confirmed_chat_seq_num(&self) -> Option<u64> {
+    pub fn get_confirmed_chat_seq_num(&self) -> Option<u64> {
         if self.game_info.is_none() {
             return None;
         }
@@ -143,7 +121,7 @@ impl Player {
 
     // Allow dead_code for unit testing
     #[allow(dead_code)]
-    fn has_chatted(&self) -> bool {
+    pub fn has_chatted(&self) -> bool {
         if self.game_info.is_none() {
             return false;
         }
@@ -157,37 +135,37 @@ impl Player {
 }
 
 #[derive(PartialEq, Debug, Clone)]
-struct ServerChatMessage {
-    seq_num:     u64,     // sequence number
-    player_id:   PlayerID,
-    player_name: String,
-    message:     String,
-    timestamp:   Instant,
+pub struct ServerChatMessage {
+    pub seq_num:     u64,     // sequence number
+    pub player_id:   PlayerID,
+    pub player_name: String,
+    pub message:     String,
+    pub timestamp:   Instant,
 }
 
 #[derive(Clone, PartialEq)]
-struct Room {
-    room_id: RoomID,
-    name:         String,
-    player_ids:   Vec<PlayerID>,
-    game_running: bool,
-    universe:     u64,    // Temp until we integrate
-    latest_seq_num: u64,
-    messages:     VecDeque<ServerChatMessage>    // Front == Oldest, Back == Newest
+pub struct Room {
+    pub room_id: RoomID,
+    pub name:         String,
+    pub player_ids:   Vec<PlayerID>,
+    pub game_running: bool,
+    pub universe:     u64,    // Temp until we integrate
+    pub latest_seq_num: u64,
+    pub messages:     VecDeque<ServerChatMessage>    // Front == Oldest, Back == Newest
 }
 
-struct ServerState {
-    tick:           usize,
-    players:        HashMap<PlayerID, Player>,
-    player_map:     HashMap<String, PlayerID>,  // map cookie to player ID
-    rooms:          HashMap<RoomID, Room>,
-    room_map:       HashMap<String, RoomID>,    // map room name to room ID
-    network_map:    HashMap<PlayerID, NetworkManager>,  // map Player ID to Player's network data
+pub struct ServerState {
+    pub tick:           usize,
+    pub players:        HashMap<PlayerID, Player>,
+    pub player_map:     HashMap<String, PlayerID>,  // map cookie to player ID
+    pub rooms:          HashMap<RoomID, Room>,
+    pub room_map:       HashMap<String, RoomID>,    // map room name to room ID
+    pub network_map:    HashMap<PlayerID, NetworkManager>,  // map Player ID to Player's network data
 }
 
 //////////////// Utilities ///////////////////////
 
-fn new_cookie() -> String {
+pub fn new_cookie() -> String {
     let mut buf = [0u8; 16];
     rand::thread_rng().fill_bytes(&mut buf);
     base64::encode(&buf)
@@ -204,7 +182,7 @@ fn new_cookie() -> String {
 *  | timestamp | rand_salt |
 *  |___________|___________|
 */
-fn new_uuid() -> u64 {
+pub fn new_uuid() -> u64 {
     let hash: u64;
 
     let mut timestamp: u64 = time::Instant::now().elapsed().as_secs().into();
@@ -217,15 +195,15 @@ fn new_uuid() -> u64 {
     hash
 }
 
-fn validate_client_version(client_version: String) -> bool {
-    let server_version = net::get_version();
+pub fn validate_client_version(client_version: String) -> bool {
+    let server_version = get_version();
 
     // Client cannot be newer than server
     server_version >= Version::parse(&client_version)
 }
 
 impl ServerChatMessage {
-    fn new(id: PlayerID, name: String, msg: String, seq_num: u64) -> Self {
+    pub fn new(id: PlayerID, name: String, msg: String, seq_num: u64) -> Self {
         ServerChatMessage {
             player_id: id,
             player_name: name,
@@ -239,7 +217,7 @@ impl ServerChatMessage {
 impl Room {
     /// Instantiates a `Room` with the provided `name` and adds
     /// the players (via `player_ids`) immediately to it.
-    fn new(name: String, player_ids: Vec<PlayerID>) -> Self {
+    pub fn new(name: String, player_ids: Vec<PlayerID>) -> Self {
         Room {
             room_id: RoomID(new_uuid()),
             name: name,
@@ -253,7 +231,7 @@ impl Room {
 
     /// The room message queue cannot exceed `MAX_NUM_CHAT_MESSAGES` so we
     /// will dequeue the oldest messages until we are within limits.
-    fn discard_older_messages(&mut self) {
+    pub fn discard_older_messages(&mut self) {
         let queue_size = self.messages.len();
         if queue_size >= MAX_NUM_CHAT_MESSAGES {
             for _ in 0..(queue_size-MAX_NUM_CHAT_MESSAGES+1) {
@@ -262,35 +240,35 @@ impl Room {
         }
     }
 
-    fn has_players(&mut self) -> bool {
+    pub fn has_players(&mut self) -> bool {
         !self.player_ids.is_empty()
     }
 
     /// Increments the room's latest sequence number
-    fn increment_seq_num(&mut self) -> u64 {
+    pub fn increment_seq_num(&mut self) -> u64 {
         self.latest_seq_num += 1;
         self.latest_seq_num
     }
 
     /// Adds a new message to the room message queue
-    fn add_message(&mut self, new_message: ServerChatMessage) {
+    pub fn add_message(&mut self, new_message: ServerChatMessage) {
         self.messages.push_back(new_message);
     }
 
     /// Gets the oldest message in the room message queue
-    fn get_oldest_msg(&self) -> Option<&ServerChatMessage> {
+    pub fn get_oldest_msg(&self) -> Option<&ServerChatMessage> {
         return self.messages.front();
     }
 
     /// Gets the newest message in the room message queue
-    fn get_newest_msg(&self) -> Option<&ServerChatMessage> {
+    pub fn get_newest_msg(&self) -> Option<&ServerChatMessage> {
         return self.messages.back();
     }
 
     /// This function retrieves the number of messages that have
     /// already been acknowledged by the client. One use of this is
     /// to only send unread messages.
-    fn get_message_skip_count(&self, chat_msg_seq_num: u64) -> u64 {
+    pub fn get_message_skip_count(&self, chat_msg_seq_num: u64) -> u64 {
         let opt_newest_msg = self.get_newest_msg();
         if opt_newest_msg.is_none() {
             return 0;
@@ -318,7 +296,7 @@ impl Room {
     }
 
     /// Send a message to all players in room notifying that an event took place.
-    fn broadcast(&mut self, event: String) {
+    pub fn broadcast(&mut self, event: String) {
         self.discard_older_messages();
         let seq_num = self.increment_seq_num();
         self.add_message(ServerChatMessage::new(SERVER_ID, "Server".to_owned(), event, seq_num));
@@ -327,7 +305,7 @@ impl Room {
 
 impl ServerState {
 
-    fn get_player(&self, player_id: PlayerID) -> &Player {
+    pub fn get_player(&self, player_id: PlayerID) -> &Player {
         let opt_player = self.players.get(&player_id);
 
         if opt_player.is_none() {
@@ -337,7 +315,7 @@ impl ServerState {
         opt_player.unwrap()
     }
 
-    fn get_player_mut(&mut self, player_id: PlayerID) -> &mut Player {
+    pub fn get_player_mut(&mut self, player_id: PlayerID) -> &mut Player {
         let opt_player = self.players.get_mut(&player_id);
 
         if opt_player.is_none() {
@@ -347,7 +325,7 @@ impl ServerState {
         opt_player.unwrap()
     }
 
-    fn get_room_id(&self, player_id: PlayerID) -> Option<RoomID> {
+    pub fn get_room_id(&self, player_id: PlayerID) -> Option<RoomID> {
         let player = self.get_player(player_id);
         if player.game_info == None {
             return None;
@@ -356,7 +334,7 @@ impl ServerState {
         Some(player.game_info.as_ref().unwrap().room_id)  // unwrap ok because of test above
     }
 
-    fn get_room_mut(&mut self, player_id: PlayerID) -> Option<&mut Room> {
+    pub fn get_room_mut(&mut self, player_id: PlayerID) -> Option<&mut Room> {
         let opt_room_id = self.get_room_id(player_id);
 
         if opt_room_id.is_none() {
@@ -365,7 +343,7 @@ impl ServerState {
         self.rooms.get_mut(&opt_room_id.unwrap())
     }
 
-    fn get_room(&self, player_id: PlayerID) -> Option<&Room> {
+    pub fn get_room(&self, player_id: PlayerID) -> Option<&Room> {
         let opt_room_id = self.get_room_id(player_id);
 
         if opt_room_id.is_none() {
@@ -374,7 +352,7 @@ impl ServerState {
         self.rooms.get(&opt_room_id.unwrap())
     }
 
-    fn list_players(&self, player_id: PlayerID) -> ResponseCode {
+    pub fn list_players(&self, player_id: PlayerID) -> ResponseCode {
         let opt_room = self.get_room(player_id);
         if opt_room.is_none() {
             return ResponseCode::BadRequest(Some("cannot list players because in lobby.".to_owned()));
@@ -391,7 +369,7 @@ impl ServerState {
         return ResponseCode::PlayerList(players);
     }
 
-    fn handle_chat_message(&mut self, player_id: PlayerID, msg: String) -> ResponseCode {
+    pub fn handle_chat_message(&mut self, player_id: PlayerID, msg: String) -> ResponseCode {
         let player_in_game = self.is_player_in_game(player_id);
 
         if !player_in_game {
@@ -420,7 +398,7 @@ impl ServerState {
         return ResponseCode::OK;
     }
 
-    fn list_rooms(&mut self) -> ResponseCode {
+    pub fn list_rooms(&mut self) -> ResponseCode {
         let mut rooms = vec![];
         self.rooms.values().for_each(|gs| {
             rooms.push((gs.name.clone(), gs.player_ids.len() as u64, gs.game_running));
@@ -428,14 +406,14 @@ impl ServerState {
         ResponseCode::RoomList(rooms)
     }
 
-    fn new_room(&mut self, name: String) {
+    pub fn new_room(&mut self, name: String) {
         let room = Room::new(name.clone(), vec![]);
 
         self.room_map.insert(name, room.room_id);
         self.rooms.insert(room.room_id, room);
     }
 
-    fn create_new_room(&mut self, opt_player_id: Option<PlayerID>, room_name: String) -> ResponseCode {
+    pub fn create_new_room(&mut self, opt_player_id: Option<PlayerID>, room_name: String) -> ResponseCode {
         // validate length
         if room_name.len() > MAX_ROOM_NAME {
             return ResponseCode::BadRequest(Some(format!("room name too long; max {} characters",
@@ -458,7 +436,7 @@ impl ServerState {
         }
     }
 
-    fn join_room(&mut self, player_id: PlayerID, room_name: String) -> ResponseCode {
+    pub fn join_room(&mut self, player_id: PlayerID, room_name: String) -> ResponseCode {
         let already_playing = self.is_player_in_game(player_id);
         if already_playing {
             return ResponseCode::BadRequest(Some("cannot join game because in-game".to_owned()));
@@ -480,7 +458,7 @@ impl ServerState {
         return ResponseCode::BadRequest(Some(format!("no room named {:?}", room_name)));
     }
 
-    fn leave_room(&mut self, player_id: PlayerID) -> ResponseCode {
+    pub fn leave_room(&mut self, player_id: PlayerID) -> ResponseCode {
         let already_playing = self.is_player_in_game(player_id);
         if !already_playing {
             return ResponseCode::BadRequest(Some("cannot leave game because in lobby".to_owned()));
@@ -502,12 +480,12 @@ impl ServerState {
         return ResponseCode::LeaveRoom;
     }
 
-    fn remove_player(&mut self, player_id: PlayerID, player_cookie: String) {
+    pub fn remove_player(&mut self, player_id: PlayerID, player_cookie: String) {
         self.player_map.remove(&player_cookie);
         self.players.remove(&player_id);
     }
 
-    fn handle_disconnect(&mut self, player_id: PlayerID) -> ResponseCode {
+    pub fn handle_disconnect(&mut self, player_id: PlayerID) -> ResponseCode {
         let (player_name, player_cookie) = {
             let player = self.get_player(player_id);
             let name = player.name.to_owned();
@@ -529,7 +507,7 @@ impl ServerState {
     }
 
     // not used for connect
-    fn process_request_action(&mut self, player_id: PlayerID, action: RequestAction) -> ResponseCode {
+    pub fn process_request_action(&mut self, player_id: PlayerID, action: RequestAction) -> ResponseCode {
         match action {
             RequestAction::Disconnect      => {
                 return self.handle_disconnect(player_id);
@@ -566,12 +544,12 @@ impl ServerState {
         }
     }
 
-    fn is_player_in_game(&self, player_id: PlayerID) -> bool {
+    pub fn is_player_in_game(&self, player_id: PlayerID) -> bool {
         let player: Option<&Player> = self.players.get(&player_id);
         player.is_some() && player.unwrap().game_info.is_some()
     }
 
-    fn is_unique_player_name(&self, name: &str) -> bool {
+    pub fn is_unique_player_name(&self, name: &str) -> bool {
         for ref player in self.players.values() {
             if player.name == name {
                 return false;
@@ -583,7 +561,7 @@ impl ServerState {
     // Request_ack contains the last processed sequence number. If one arrives older (less than)
     // than the last processed, it must be rejected.
     // FIXME Does not handle wrapped sequence number case yet.
-    fn is_previously_processed_packet(&mut self, player_id: PlayerID, sequence: u64) -> bool {
+    pub fn is_previously_processed_packet(&mut self, player_id: PlayerID, sequence: u64) -> bool {
         let player: &Player = self.get_player(player_id);
         if let Some(request_ack) = player.request_ack {
             if sequence <= request_ack {
@@ -593,7 +571,7 @@ impl ServerState {
         false
     }
 
-    fn get_player_id_by_cookie(&self, cookie: &str) -> Option<PlayerID> {
+    pub fn get_player_id_by_cookie(&self, cookie: &str) -> Option<PlayerID> {
         match self.player_map.get(cookie) {
             Some(player_id) => Some(*player_id),
             None => None
@@ -602,7 +580,7 @@ impl ServerState {
 
     /// Returns true if the packet already exists in the queue, otherwise it will return false, and
     /// will be added in sequence_number order.
-    fn add_packet_to_queue(&mut self, player_id: PlayerID, packet: Packet) -> bool {
+    pub fn add_packet_to_queue(&mut self, player_id: PlayerID, packet: Packet) -> bool {
         // Unwrap should be safe since a player ID was already found.
         let network: &mut NetworkManager = self.network_map.get_mut(&player_id).unwrap();
         let already_exists = network.rx_packets.buffer_item(packet);
@@ -610,7 +588,7 @@ impl ServerState {
     }
 
     /// Checks to see if the incoming packet is immediately processable
-    fn can_process_packet(&mut self, player_id: PlayerID, sequence_number: u64) -> bool {
+    pub fn can_process_packet(&mut self, player_id: PlayerID, sequence_number: u64) -> bool {
         let player: &mut Player = self.get_player_mut(player_id);
         if let Some(ack) = player.request_ack {
             trace!("[CAN PROCESS?] Ack: {} Sqn: {}", ack, sequence_number);
@@ -624,7 +602,7 @@ impl ServerState {
 
     /// Processes a player's request action for all non Connect requests. If necessary, a response is buffered
     /// for later transmission
-    fn process_player_request_action(&mut self, player_id: PlayerID, action: RequestAction) -> Result<Option<Packet>, Box<Error>> {
+    pub fn process_player_request_action(&mut self, player_id: PlayerID, action: RequestAction) -> Result<Option<Packet>, Box<Error>> {
         match action {
             RequestAction::Connect{..} => unreachable!(),
             _ => {
@@ -644,7 +622,7 @@ impl ServerState {
     }
 
     /// Determine how many contiguous packets are processable and process their requests for the given player.
-    fn process_queued_rx_packets(&mut self, player_id: PlayerID) {
+    pub fn process_queued_rx_packets(&mut self, player_id: PlayerID) {
         // If we can, start popping off the RX queue and handle contiguous packets immediately
         let mut dequeue_count = 0;
 
@@ -695,13 +673,13 @@ impl ServerState {
         }
     }
 
-    fn process_player_buffered_packets(&mut self, players_to_update: &Vec<PlayerID>) {
+    pub fn process_player_buffered_packets(&mut self, players_to_update: &Vec<PlayerID>) {
         for player_id in players_to_update {
             self.process_queued_rx_packets(*player_id);
         }
     }
 
-    fn process_buffered_packets_in_lobby(&mut self){
+    pub fn process_buffered_packets_in_lobby(&mut self){
         let mut players_to_update: Vec<PlayerID> = vec![];
 
         if self.players.len() == 0 {
@@ -721,7 +699,7 @@ impl ServerState {
         }
     }
 
-    fn resend_expired_tx_packets(&mut self, udp_tx: &mpsc::UnboundedSender<(SocketAddr, Packet)>) {
+    pub fn resend_expired_tx_packets(&mut self, udp_tx: &mpsc::UnboundedSender<(SocketAddr, Packet)>) {
         let mut players_to_update: Vec<PlayerID> = vec![];
 
         if self.players.len() == 0 {
@@ -756,7 +734,7 @@ impl ServerState {
         }
     }
 
-    fn process_buffered_packets_in_rooms(&mut self) {
+    pub fn process_buffered_packets_in_rooms(&mut self) {
         let mut players_to_update: Vec<PlayerID> = vec![];
 
         if self.rooms.len() == 0 {
@@ -785,7 +763,7 @@ impl ServerState {
     }
 
     /// Clear out the transmission queue of any packets the client has acknowledged
-    fn clear_transmission_queue_on_ack(&mut self, player_id: PlayerID, response_ack: Option<u64>) {
+    pub fn clear_transmission_queue_on_ack(&mut self, player_id: PlayerID, response_ack: Option<u64>) {
         if let Some(response_ack) = response_ack {
             if let Some(ref mut player_network) = self.network_map.get_mut(&player_id) {
                 let mut removal_count = 0;
@@ -816,7 +794,7 @@ impl ServerState {
     ///  3. Client should notified if version requires updating
     ///  4. Ignore if already received or processed
     /// Always returns either Ok(Some(Packet::Response{...})), Ok(None), or error.
-    fn decode_packet(&mut self, addr: SocketAddr, packet: Packet) -> Result<Option<Packet>, Box<Error>> {
+    pub fn decode_packet(&mut self, addr: SocketAddr, packet: Packet) -> Result<Option<Packet>, Box<Error>> {
         match packet.clone() {
             _pkt @ Packet::Response{..} | _pkt @ Packet::Update{..} => {
                 return Err(Box::new(io::Error::new(ErrorKind::InvalidData, "invalid packet type")));
@@ -921,7 +899,7 @@ impl ServerState {
         }
     }
 
-    fn prepare_response(&mut self, player_id: PlayerID, action: RequestAction) -> Option<Packet> {
+    pub fn prepare_response(&mut self, player_id: PlayerID, action: RequestAction) -> Option<Packet> {
         let response_code = self.process_request_action(player_id, action.clone());
 
         let (sequence, request_ack);//= (0, None);
@@ -957,7 +935,7 @@ impl ServerState {
         })
     }
 
-    fn handle_new_connection(&mut self, name: String, addr: SocketAddr) -> Packet {
+    pub fn handle_new_connection(&mut self, name: String, addr: SocketAddr) -> Packet {
         if self.is_unique_player_name(&name) {
             let player = self.add_new_player(name.clone(), addr.clone());
             let cookie = player.cookie.clone();
@@ -966,7 +944,7 @@ impl ServerState {
             let response = Packet::Response{
                 sequence:    0,
                 request_ack: Some(0), // Should start at seq_num 0 unless client's network state was not properly reset
-                code:        ResponseCode::LoggedIn(cookie, net::VERSION.to_owned()),
+                code:        ResponseCode::LoggedIn(cookie, VERSION.to_owned()),
             };
             return response;
         } else {
@@ -981,7 +959,7 @@ impl ServerState {
     }
 
     // Right now we'll be constructing all client Update packets for _every_ room.
-    fn construct_client_updates(&mut self) -> Result<Option<Vec<(SocketAddr, Packet)>>, Box<Error>> {
+    pub fn construct_client_updates(&mut self) -> Result<Option<Vec<(SocketAddr, Packet)>>, Box<Error>> {
         let mut client_updates: Vec<(SocketAddr, Packet)> = vec![];
 
         if self.rooms.len() == 0 {
@@ -1031,7 +1009,7 @@ impl ServerState {
 
     /// Creates a vector of messages that the provided Player has not yet acknowledged.
     /// Exits early if the player is already caught up.
-    fn collect_unacknowledged_messages(&self, room: &Room, player: &Player) -> Option<Vec<BroadcastChatMessage>> {
+    pub fn collect_unacknowledged_messages(&self, room: &Room, player: &Player) -> Option<Vec<BroadcastChatMessage>> {
         // Only send what a player has not yet seen
         let raw_unsent_messages: VecDeque<ServerChatMessage>;
         match player.get_confirmed_chat_seq_num() {
@@ -1073,7 +1051,7 @@ impl ServerState {
         return Some(unsent_messages);
     }
 
-    fn expire_old_messages_in_all_rooms(&mut self) {
+    pub fn expire_old_messages_in_all_rooms(&mut self) {
         if self.rooms.len() != 0 {
             let current_timestamp = time::Instant::now();
             for room in self.rooms.values_mut() {
@@ -1084,7 +1062,7 @@ impl ServerState {
         }
     }
 
-    fn add_new_player(&mut self, name: String, addr: SocketAddr) -> &mut Player {
+    pub fn add_new_player(&mut self, name: String, addr: SocketAddr) -> &mut Player {
         let cookie = new_cookie();
         let player_id = PlayerID(new_uuid());
         let player = Player {
@@ -1110,11 +1088,11 @@ impl ServerState {
         player
     }
 
-    fn remove_timed_out_clients(&mut self) {
+    pub fn remove_timed_out_clients(&mut self) {
         let mut timed_out_players: Vec<PlayerID> = vec![];
 
         for (p_id, p) in self.players.iter() {
-            if net::has_connection_timed_out(p.heartbeat) {
+            if has_connection_timed_out(p.heartbeat) {
                 info!("Player({}) has timed out", p.cookie);
                 timed_out_players.push(*p_id);
             }
@@ -1126,7 +1104,7 @@ impl ServerState {
 
     }
 
-    fn new() -> Self {
+    pub fn new() -> Self {
         ServerState {
             tick:       0,
             players:    HashMap::<PlayerID, Player>::new(),
@@ -1134,1549 +1112,6 @@ impl ServerState {
             player_map: HashMap::<String, PlayerID>::new(),
             room_map:   HashMap::<String, RoomID>::new(),
             network_map: HashMap::<PlayerID, NetworkManager>::new(),
-        }
-    }
-}
-
-//////////////// Event Handling /////////////////
-enum Event {
-    TickEvent,
-    Request((SocketAddr, Option<Packet>)),
-    NetworkEvent,
-    HeartBeat,
-//    Notify((SocketAddr, Option<Packet>)),
-}
-
-//////////////////// Main /////////////////////
-fn main() {
-    env_logger::Builder::new()
-    .format(|buf, record| {
-        writeln!(buf,
-            "{} [{:5}] - {}",
-            Local::now().format("%a %Y-%m-%d %H:%M:%S%.6f"),
-            record.level(),
-            record.args(),
-        )
-    })
-    .filter(None, LevelFilter::Trace)
-    .filter(Some("futures"), LevelFilter::Off)
-    .filter(Some("tokio_core"), LevelFilter::Off)
-    .filter(Some("tokio_reactor"), LevelFilter::Off)
-    .init();
-
-    let mut core = Core::new().unwrap();
-    let handle = core.handle();
-
-    let (tx, rx) = mpsc::unbounded();
-
-    let udp = net::bind(&handle, None, None)
-        .unwrap_or_else(|e| {
-            error!("Error while trying to bind UDP socket: {:?}", e);
-            exit(1);
-        });
-
-    let (udp_sink, udp_stream) = udp.framed(LineCodec).split();
-
-    let initial_server_state = ServerState::new();
-
-    let iter_stream = stream::iter_ok::<_, io::Error>(iter::repeat( () ));
-    let tick_stream = iter_stream.and_then(|_| {
-        let timeout = Timeout::new(Duration::from_millis(TICK_INTERVAL_IN_MS), &handle).unwrap();
-        timeout.and_then(move |_| {
-            ok(Event::TickEvent)
-        })
-    }).map_err(|_| ());
-
-    let packet_stream = udp_stream
-        .filter(|&(_, ref opt_packet)| {
-            *opt_packet != None
-        })
-        .map(|packet_tuple| {
-            Event::Request(packet_tuple)
-        })
-        .map_err(|_| ());
-
-    let network_stream = stream::iter_ok::<_, io::Error>(iter::repeat( () ));
-    let network_stream = network_stream.and_then(|_| {
-        let timeout = Timeout::new(Duration::from_millis(NETWORK_INTERVAL_IN_MS), &handle).unwrap();
-        timeout.and_then(move |_| {
-            ok(Event::NetworkEvent)
-        })
-    }).map_err(|_| ());
-
-    let heartbeat_stream = stream::iter_ok::<_, io::Error>(iter::repeat( () ));
-    let heartbeat_stream = heartbeat_stream.and_then(|_| {
-        let timeout = Timeout::new(Duration::from_millis(HEARTBEAT_INTERVAL_IN_MS), &handle).unwrap();
-        timeout.and_then(move |_| {
-            ok(Event::HeartBeat)
-        })
-    }).map_err(|_| ());
-
-    let server_fut = tick_stream
-        .select(packet_stream)
-        .select(network_stream)
-        .select(heartbeat_stream)
-        .fold(initial_server_state, move |mut server_state: ServerState, event: Event | {
-            match event {
-                Event::Request(packet_tuple) => {
-                     // With the above filter, `packet` should never be None
-                    let (addr, opt_packet) = packet_tuple;
-
-                    // Decode incoming and send a Response to the Requester
-                    if let Some(packet) = opt_packet {
-                        let decode_result = server_state.decode_packet(addr, packet.clone());
-                        if decode_result.is_ok() {
-                            let opt_response_packet = decode_result.unwrap();
-
-                            if let Some(response_packet) = opt_response_packet {
-                                let response = (addr.clone(), response_packet);
-                                netwayste_send!(tx, response, ("[EVENT::REQUEST] Immediate response failed."));
-                            }
-                        } else {
-                            let err = decode_result.unwrap_err();
-                            error!("Decoding packet failed, from {:?}: {:?}", addr, err);
-                        }
-                    }
-                }
-
-                Event::TickEvent => {
-                    server_state.expire_old_messages_in_all_rooms();
-                    let client_update_packets_result = server_state.construct_client_updates();
-                    if client_update_packets_result.is_ok() {
-                        let opt_update_packets = client_update_packets_result.unwrap();
-
-                        if let Some(update_packets) = opt_update_packets {
-                            for update in update_packets {
-                                netwayste_send!(tx, update, ("[EVENT::TICK] Could not send client update."));
-                            }
-                        }
-                    }
-
-                    /*
-                    for x in server_state.network_map.values() {
-                        trace!("\n\n\nNETWORK QUEUE CAPACITIES\n-----------------------\nTX: {}\nRX: {}\n\n\n", x.tx_packets.as_queue_type().capacity(), x.rx_packets.as_queue_type().capacity());
-                    }
-                    */
-
-                    server_state.remove_timed_out_clients();
-                    server_state.tick  = 1usize.wrapping_add(server_state.tick);
-                }
-
-                Event::NetworkEvent => {
-                    // Process players in rooms
-                    server_state.process_buffered_packets_in_rooms();
-
-                    // Process players in lobby
-                    server_state.process_buffered_packets_in_lobby();
-
-                    server_state.resend_expired_tx_packets(&tx);
-                }
-
-                Event::HeartBeat => {
-                    for player in server_state.players.values() {
-                        let keep_alive = Packet::Response {
-                            sequence: 0,
-                            request_ack: None,
-                            code: ResponseCode::KeepAlive
-                        };
-                        netwayste_send!(tx, (player.addr, keep_alive), ("[EVENT::HEARTBEAT] Could not send to Player: {:?}", player));
-                    }
-                }
-            }
-
-            // return the updated client for the next iteration
-            ok(server_state)
-        })
-        .map(|_| ())
-        .map_err(|_| ());
-
-    let sink_fut = rx.fold(udp_sink, |udp_sink, outgoing_item| {
-            let udp_sink = udp_sink.send(outgoing_item).map_err(|_| ());    // this method flushes (if too slow, use send_all)
-            udp_sink
-        }).map(|_| ()).map_err(|_| ());
-
-    let combined_fut = server_fut.map(|_| ())
-        .select(sink_fut)
-        .map(|_| ());   // wait for either server_fut or sink_fut to complete
-
-    drop(core.run(combined_fut));
-}
-
-#[cfg(test)]
-mod test {
-    use super::*;
-    use proptest::strategy::*;
-
-    fn fake_socket_addr() -> SocketAddr {
-        use std::net::{IpAddr, Ipv4Addr};
-        SocketAddr::new(IpAddr::V4(Ipv4Addr::new(1, 2, 3, 4)), 5678)
-    }
-
-    #[test]
-    fn list_players_player_shows_up_in_player_list() {
-        let mut server = ServerState::new();
-        let room_name = "some name";
-        // make a new room
-        server.create_new_room(None, String::from(room_name));
-
-        let (player_id, player_name) = {
-            let p: &mut Player = server.add_new_player(String::from("some name"), fake_socket_addr());
-
-            (p.player_id, p.name.clone())
-        };
-        // make the player join the room
-        {
-            server.join_room(player_id, String::from(room_name));
-        }
-        let resp_code: ResponseCode = server.list_players(player_id);
-        match resp_code {
-            ResponseCode::PlayerList(players) => {
-                assert_eq!(players.len(), 1);
-                assert_eq!(*players.first().unwrap(), player_name);
-            }
-            resp_code @ _ => panic!("Unexpected response code: {:?}", resp_code)
-        }
-    }
-
-    #[test]
-    fn has_chatted_player_did_not_chat_on_join() {
-        let mut server = ServerState::new();
-        let room_name = "some name";
-        // make a new room
-        server.create_new_room(None, String::from(room_name));
-        let player_id = {
-            let p: &mut Player = server.add_new_player(String::from("some name"), fake_socket_addr());
-            p.player_id
-        };
-        // make the player join the room
-        {
-            server.join_room(player_id, String::from(room_name));
-        }
-        let player = server.get_player(player_id);
-        assert_eq!(player.has_chatted(), false);
-    }
-
-    #[test]
-    fn get_confirmed_chat_seq_num_server_tracks_players_chat_updates() {
-        let mut server = ServerState::new();
-        let room_name = "some name";
-        // make a new room
-        server.create_new_room(None, String::from(room_name));
-
-        let (player_id, player_cookie) = {
-            let p: &mut Player = server.add_new_player(String::from("some name"), fake_socket_addr());
-
-            (p.player_id, p.cookie.clone())
-        };
-        // make the player join the room
-        {
-            server.join_room(player_id, String::from(room_name));
-        }
-
-        // A chat-less player now has something to to say
-        server.decode_packet(fake_socket_addr(), Packet::UpdateReply {
-            cookie: player_cookie.clone(),
-            last_chat_seq: Some(1),
-            last_game_update_seq: None,
-            last_gen: None
-        }).unwrap();
-
-        {
-            let player = server.get_player(player_id);
-            assert_eq!(player.get_confirmed_chat_seq_num(), Some(1));
-        }
-
-        // Older messages are ignored
-        server.decode_packet(fake_socket_addr(), Packet::UpdateReply {
-            cookie: player_cookie.clone(),
-            last_chat_seq: Some(0),
-            last_game_update_seq: None,
-            last_gen: None
-        }).unwrap();
-
-        {
-            let player = server.get_player(player_id);
-            assert_eq!(player.get_confirmed_chat_seq_num(), Some(1));
-        }
-
-        // So are absent messages
-        server.decode_packet(fake_socket_addr(), Packet::UpdateReply {
-            cookie: player_cookie,
-            last_chat_seq: None,
-            last_game_update_seq: None,
-            last_gen: None
-        }).unwrap();
-
-        {
-            let player = server.get_player(player_id);
-            assert_eq!(player.get_confirmed_chat_seq_num(), Some(1));
-        }
-    }
-
-    #[test]
-    fn get_message_skip_count_player_acked_messages_are_not_included_in_skip_count() {
-        let mut server = ServerState::new();
-        let room_name = "some name";
-
-        server.create_new_room(None, String::from(room_name));
-
-        let (player_id, _) = {
-            let p: &mut Player = server.add_new_player(String::from("some name"), fake_socket_addr());
-
-            (p.player_id, p.cookie.clone())
-        };
-        // make the player join the room
-        // Give it a single message
-        {
-            server.join_room(player_id, String::from(room_name));
-            server.handle_chat_message(player_id, "ChatMessage".to_owned());
-        }
-
-        {
-            let room: &Room = server.get_room(player_id).unwrap();
-            // The check below does not affect any player acknowledgement as we are not
-            // involving the player yet. This is a simple test to ensure that if a chat
-            // message decoded from a would-be player was less than the latest chat message,
-            // we handle it properly by not skipping any.
-            assert_eq!(room.get_message_skip_count(0), 0);
-        }
-
-        let number_of_messages = 6;
-        for _ in 1..number_of_messages {
-            server.handle_chat_message(player_id, "ChatMessage".to_owned());
-        }
-
-        {
-            //let player = server.get_player_mut(player_id);
-            let player = server.get_player_mut(player_id);
-            // player has not acknowledged any yet
-            #[should_panic]
-            assert_eq!(player.get_confirmed_chat_seq_num(), None);
-        }
-
-        // player acknowledged four of the six
-        let acked_message_count = {
-            let player = server.get_player_mut(player_id);
-            player.update_chat_seq_num(Some(4));
-
-            player.get_confirmed_chat_seq_num().unwrap()
-        };
-        {
-            let room: &Room = server.get_room(player_id).unwrap();
-            assert_eq!(room.get_message_skip_count(acked_message_count), acked_message_count);
-        }
-
-        // player acknowledged all six
-        let acked_message_count = {
-            let player = server.get_player_mut(player_id);
-            player.update_chat_seq_num(Some(6));
-
-            player.get_confirmed_chat_seq_num().unwrap()
-        };
-        {
-            let room: &Room = server.get_room(player_id).unwrap();
-            assert_eq!(room.get_message_skip_count(acked_message_count), acked_message_count);
-        }
-    }
-
-    #[test]
-    // Send fifteen messages, but only leave nine unacknowledged, while wrapping on the sequence number
-    fn get_message_skip_count_player_acked_messages_are_not_included_in_skip_count_wrapped_case() {
-        let mut server = ServerState::new();
-        let room_name = "some name";
-
-        server.create_new_room(None, String::from(room_name));
-
-        let player_id = {
-            let p: &mut Player = server.add_new_player(String::from("some name"), fake_socket_addr());
-
-            p.player_id
-        };
-        {
-            server.join_room(player_id, String::from(room_name));
-        }
-
-        // Picking a value slightly less than max of u64
-        let start_seq_num = u64::max_value() - 6;
-        // First pass, add messages with sequence numbers through the max of u64
-        for seq_num in start_seq_num..u64::max_value() {
-            let room: &mut Room = server.get_room_mut(player_id).unwrap();
-            room.add_message(ServerChatMessage::new(player_id, String::from("some name"), String::from("some msg"), seq_num));
-        }
-        // Second pass, from wrap-point, `0`, eight times
-        for seq_num in 0..8 {
-            let room: &mut Room = server.get_room_mut(player_id).unwrap();
-            room.add_message(ServerChatMessage::new(player_id, String::from("some name"), String::from("some msg"), seq_num));
-        }
-
-        let acked_message_count = {
-            // Ack up until 0xFFFFFFFFFFFFFFFD
-            let player = server.get_player_mut(player_id);
-            player.update_chat_seq_num(Some(start_seq_num + 4));
-
-            player.get_confirmed_chat_seq_num().unwrap()
-        };
-        {
-            let room: &Room = server.get_room(player_id).unwrap();
-            // Fifteen total messages sent.
-            // 2 unacked which are less than u64::max_value()
-            // 8 unacked which are after the numerical wrap
-            let unacked_count = 15 - (8 + 2);
-            assert_eq!(room.get_message_skip_count(acked_message_count), unacked_count);
-        }
-    }
-
-    #[test]
-    fn collect_unacknowledged_messages_a_rooms_unacknowledged_chat_messages_are_collected_for_their_player() {
-        let mut server = ServerState::new();
-        let room_name = "some name";
-
-        server.create_new_room(None, String::from(room_name));
-
-        let player_id = {
-            let p: &mut Player = server.add_new_player(String::from("some name"), fake_socket_addr());
-
-            p.player_id
-        };
-        {
-            server.join_room(player_id, String::from(room_name));
-        }
-
-        {
-            // Room has no messages, None to send to player
-            let room = server.get_room(player_id).unwrap();
-            let player = server.get_player(player_id);
-            let messages = server.collect_unacknowledged_messages(room, player);
-            assert_eq!(messages, None);
-        }
-
-        {
-            let room: &mut Room = server.get_room_mut(player_id).unwrap();
-            room.add_message(ServerChatMessage::new(player_id, String::from("some name"), String::from("some msg"), 1));
-        }
-        {
-            // Room has a message, player has yet to ack it
-            let room = server.get_room(player_id).unwrap();
-            let player = server.get_player(player_id);
-            let messages = server.collect_unacknowledged_messages(room, player);
-            assert_eq!(messages.is_some(), true);
-            assert_eq!(messages.unwrap().len(), 1);
-        }
-
-        {
-            let player = server.get_player_mut(player_id);
-            player.update_chat_seq_num(Some(1));
-        }
-        {
-            // Room has a message, player acked, None
-            let room = server.get_room(player_id).unwrap();
-            let player = server.get_player(player_id);
-            let messages = server.collect_unacknowledged_messages(room, player);
-            assert_eq!(messages, None);
-        }
-    }
-
-    #[test]
-    fn collect_unacknowledged_messages_an_active_room_which_expired_all_messages_returns_none()
-    {
-        let mut server = ServerState::new();
-        let room_name = "some name";
-
-        server.create_new_room(None, String::from(room_name));
-
-        let player_id = {
-            let p: &mut Player = server.add_new_player(String::from("some name"), fake_socket_addr());
-
-            p.player_id
-        };
-        {
-            server.join_room(player_id, String::from(room_name));
-        }
-
-        {
-            // Add a message to the room and then age it so it will expire
-            let room: &mut Room = server.get_room_mut(player_id).unwrap();
-            room.add_message(ServerChatMessage::new(player_id, String::from("some name"), String::from("some msg"), 1));
-
-            let message: &mut ServerChatMessage = room.messages.get_mut(0).unwrap();
-            message.timestamp = time::Instant::now() - Duration::from_secs(MAX_AGE_CHAT_MESSAGES as u64);
-        }
-        {
-            // Sanity check to ensure player gets the chat message if left unacknowledged
-            let room = server.get_room(player_id).unwrap();
-            let player = server.get_player(player_id);
-            let messages = server.collect_unacknowledged_messages(room, player);
-            assert_eq!(messages.is_some(), true);
-            assert_eq!(messages.unwrap().len(), 1);
-        }
-        {
-            let player = server.get_player_mut(player_id);
-            player.update_chat_seq_num(Some(1));
-        }
-
-        {
-            // Server drains expired messages for the room
-            server.expire_old_messages_in_all_rooms();
-        }
-        {
-            // A room that has no messages, but has player(s) who have acknowledged past messages
-            let room = server.get_room(player_id).unwrap();
-            let player = server.get_player(player_id);
-            let messages = server.collect_unacknowledged_messages(room, player);
-            assert_eq!(messages, None);
-        }
-    }
-
-    #[test]
-    fn handle_chat_message_player_not_in_game()
-    {
-        let mut server = ServerState::new();
-        let room_name = "some name";
-
-        server.create_new_room(None, room_name.to_owned());
-
-        let player_id = {
-            let p: &mut Player = server.add_new_player("some name".to_owned(), fake_socket_addr());
-
-            p.player_id
-        };
-
-        let response = server.handle_chat_message(player_id, "test msg".to_owned());
-        assert_eq!(response, ResponseCode::BadRequest(Some(format!("Player {} has not joined a game.", player_id))));
-    }
-
-    #[test]
-    fn handle_chat_message_player_in_game_one_message()
-    {
-
-        let mut server = ServerState::new();
-        let room_name = "some name";
-
-        server.create_new_room(None, room_name.to_owned());
-
-        let player_id = {
-            let p: &mut Player = server.add_new_player("some player".to_string(), fake_socket_addr());
-
-            p.player_id
-        };
-        {
-            server.join_room(player_id, room_name.to_owned());
-        }
-
-        let response = server.handle_chat_message(player_id, "test msg".to_owned());
-        assert_eq!(response, ResponseCode::OK);
-        let room: &Room = server.get_room(player_id).unwrap();
-        assert_eq!(room.messages.len(), 1);
-        assert_eq!(room.latest_seq_num, 1);
-        assert_eq!(room.get_newest_msg(), room.get_oldest_msg());
-    }
-
-    #[test]
-    fn handle_chat_message_player_in_game_many_messages()
-    {
-
-        let mut server = ServerState::new();
-        let room_name = "some name";
-
-        server.create_new_room(None, room_name.to_owned());
-
-        let player_id = {
-            let p: &mut Player = server.add_new_player("some player".to_owned(), fake_socket_addr());
-
-            p.player_id
-        };
-        {
-            server.join_room(player_id, room_name.to_owned());
-        }
-
-        let response = server.handle_chat_message(player_id, "test msg first".to_owned());
-        assert_eq!(response, ResponseCode::OK);
-        let response = server.handle_chat_message(player_id, "test msg second".to_owned());
-        assert_eq!(response, ResponseCode::OK);
-
-        let room: &Room = server.get_room(player_id).unwrap();
-        assert_eq!(room.messages.len(), 2);
-        assert_eq!(room.latest_seq_num, 2);
-    }
-
-    #[test]
-    fn create_new_room_good_case()
-    {
-        {
-            let mut server = ServerState::new();
-            let room_name = "some name".to_owned();
-
-            assert_eq!(server.create_new_room(None, room_name), ResponseCode::OK);
-        }
-        // Room name length is within bounds
-        {
-            let mut server = ServerState::new();
-            let room_name = "0123456789ABCDEF".to_owned();
-
-            assert_eq!(server.create_new_room(None, room_name), ResponseCode::OK);
-        }
-    }
-
-    #[test]
-    fn create_new_room_name_is_too_long()
-    {
-        let mut server = ServerState::new();
-        let room_name = "0123456789ABCDEF_#".to_owned();
-
-        assert_eq!(server.create_new_room(None, room_name), ResponseCode::BadRequest(Some("room name too long; max 16 characters".to_owned())));
-    }
-
-    #[test]
-    fn create_new_room_name_taken()
-    {
-        let mut server = ServerState::new();
-        let room_name = "some room".to_owned();
-        assert_eq!(server.create_new_room(None, room_name.clone()), ResponseCode::OK);
-        assert_eq!(server.create_new_room(None, room_name), ResponseCode::BadRequest(Some("room name already in use".to_owned())));
-    }
-
-    #[test]
-    fn create_new_room_player_already_in_room()
-    {
-        let mut server = ServerState::new();
-        let room_name = "some room".to_owned();
-        let other_room_name = "another room".to_owned();
-        assert_eq!(server.create_new_room(None, room_name.clone()), ResponseCode::OK);
-
-        let player_id = {
-            let p: &mut Player = server.add_new_player("some player".to_owned(), fake_socket_addr());
-
-            p.player_id
-        };
-        {
-            server.join_room(player_id, room_name.to_owned());
-        }
-
-        assert_eq!( server.create_new_room(Some(player_id), other_room_name), ResponseCode::BadRequest(Some("cannot create room because in-game".to_owned())) );
-    }
-
-    #[test]
-    fn create_new_room_join_room_good_case()
-    {
-
-        let mut server = ServerState::new();
-        let room_name = "some room".to_owned();
-        assert_eq!(server.create_new_room(None, room_name.clone()), ResponseCode::OK);
-
-        let player_id = {
-            let p: &mut Player = server.add_new_player("some player".to_owned(), fake_socket_addr());
-
-            p.player_id
-        };
-        assert_eq!(server.join_room(player_id, room_name.to_owned()), ResponseCode::JoinedRoom("some room".to_owned()));
-    }
-
-    #[test]
-    fn join_room_player_already_in_room()
-    {
-
-        let mut server = ServerState::new();
-        let room_name = "some room".to_owned();
-        assert_eq!(server.create_new_room(None, room_name.clone()), ResponseCode::OK);
-
-        let player_id = {
-            let p: &mut Player = server.add_new_player("some player".to_owned(), fake_socket_addr());
-
-            p.player_id
-        };
-        assert_eq!(server.join_room(player_id, room_name.clone()), ResponseCode::JoinedRoom("some room".to_owned()));
-        assert_eq!( server.join_room(player_id, room_name), ResponseCode::BadRequest(Some("cannot join game because in-game".to_owned())) );
-    }
-
-    #[test]
-    fn join_room_room_does_not_exist()
-    {
-
-        let mut server = ServerState::new();
-
-        let player_id = {
-            let p: &mut Player = server.add_new_player("some player".to_owned(), fake_socket_addr());
-
-            p.player_id
-        };
-        assert_eq!(server.join_room(player_id, "some room".to_owned()), ResponseCode::BadRequest(Some("no room named \"some room\"".to_owned())) );
-    }
-
-    #[test]
-    fn leave_room_good_case()
-    {
-        let mut server = ServerState::new();
-        let room_name = "some name";
-
-        server.create_new_room(None, room_name.to_owned());
-
-        let player_id = {
-            let p: &mut Player = server.add_new_player("some player".to_owned(), fake_socket_addr());
-
-            p.player_id
-        };
-        {
-            server.join_room(player_id, room_name.to_owned());
-        }
-
-        assert_eq!( server.leave_room(player_id), ResponseCode::LeaveRoom );
-
-    }
-
-    #[test]
-    fn leave_room_player_not_in_room()
-    {
-        let mut server = ServerState::new();
-        let room_name = "some room".to_owned();
-        assert_eq!(server.create_new_room(None, room_name.clone()), ResponseCode::OK);
-
-        let player_id = {
-            let p: &mut Player = server.add_new_player("some player".to_owned(), fake_socket_addr());
-
-            p.player_id
-        };
-
-        assert_eq!( server.leave_room(player_id), ResponseCode::BadRequest(Some("cannot leave game because in lobby".to_owned())) );
-    }
-
-    #[test]
-    fn leave_room_unregistered_player_id()
-    {
-        let mut server = ServerState::new();
-        let room_name = "some room".to_owned();
-        let rand_player_id = PlayerID(0x2457); //RUST
-        assert_eq!(server.create_new_room(None, room_name.clone()), ResponseCode::OK);
-
-        assert_eq!( server.leave_room(rand_player_id), ResponseCode::BadRequest(Some("cannot leave game because in lobby".to_owned())) );
-    }
-
-    #[test]
-    fn add_new_player_player_added_with_initial_sequence_number()
-    {
-        let mut server = ServerState::new();
-        let name = "some player".to_owned();
-
-        let p: &mut Player = server.add_new_player(name.clone(), fake_socket_addr());
-        assert_eq!(p.name, name);
-    }
-
-    #[test]
-    fn is_unique_player_name_yes_and_no_case()
-    {
-        let mut server = ServerState::new();
-        let name = "some player".to_owned();
-        assert_eq!(server.is_unique_player_name("some player"), true);
-
-        {
-            server.add_new_player(name.clone(), fake_socket_addr());
-        }
-        assert_eq!(server.is_unique_player_name("some player"), false);
-    }
-
-    #[test]
-    fn expire_old_messages_in_all_rooms_room_is_empty()
-    {
-        let mut server = ServerState::new();
-        let room_name = "some room";
-
-        server.create_new_room(None, room_name.to_owned().clone());
-        server.expire_old_messages_in_all_rooms();
-
-        for room in server.rooms.values() {
-            assert_eq!(room.messages.len(), 0);
-        }
-    }
-
-
-    #[test]
-    fn expire_old_messages_in_all_rooms_one_room_good_case()
-    {
-        let mut server = ServerState::new();
-        let room_name = "some room";
-
-        server.create_new_room(None, room_name.to_owned().clone());
-        let player_id: PlayerID = {
-            let player: &mut Player = server.add_new_player("some player".to_owned(), fake_socket_addr());
-            player.player_id
-        };
-
-        server.join_room(player_id, room_name.to_owned());
-
-        server.handle_chat_message(player_id, "Conwayste is such a fun game".to_owned());
-        server.handle_chat_message(player_id, "There are not loot boxes".to_owned());
-        server.handle_chat_message(player_id, "It is free!".to_owned());
-        server.handle_chat_message(player_id, "What's not to love?".to_owned());
-
-        let message_count = {
-            let room: &Room = server.get_room(player_id).unwrap();
-            room.messages.len()
-        };
-        assert_eq!(message_count, 4);
-
-        // Messages are not old enough to be expired
-        server.expire_old_messages_in_all_rooms();
-
-        for room in server.rooms.values() {
-            assert_eq!(room.messages.len(), 4);
-        }
-    }
-
-    #[test]
-    fn expire_old_messages_in_all_rooms_several_rooms_good_case()
-    {
-        let mut server = ServerState::new();
-        let room_name = "some room";
-        let room_name2 = "some room2";
-
-        server.create_new_room(None, room_name.to_owned().clone());
-        server.create_new_room(None, room_name2.to_owned().clone());
-        let player_id: PlayerID = {
-            let player: &mut Player = server.add_new_player("some player".to_owned(), fake_socket_addr());
-            player.player_id
-        };
-        let player_id2: PlayerID = {
-            let player: &mut Player = server.add_new_player("some player".to_owned(), fake_socket_addr());
-            player.player_id
-        };
-
-        server.join_room(player_id, room_name.to_owned());
-        server.join_room(player_id2, room_name2.to_owned());
-
-        server.handle_chat_message(player_id, "Conwayste is such a fun game".to_owned());
-        server.handle_chat_message(player_id, "There are not loot boxes".to_owned());
-        server.handle_chat_message(player_id2, "It is free!".to_owned());
-        server.handle_chat_message(player_id2, "What's not to love?".to_owned());
-
-        let message_count = {
-            let room: &Room = server.get_room(player_id).unwrap();
-            room.messages.len()
-        };
-        assert_eq!(message_count, 2);
-        let message_count2 = {
-            let room: &Room = server.get_room(player_id2).unwrap();
-            room.messages.len()
-        };
-        assert_eq!(message_count2, 2);
-
-        // Messages are not old enough to be expired
-        server.expire_old_messages_in_all_rooms();
-
-        for room in server.rooms.values() {
-            assert_eq!(room.messages.len(), 2);
-        }
-    }
-
-    #[test]
-    fn expire_old_messages_in_all_rooms_one_room_old_messages_are_wiped()
-    {
-        let mut server = ServerState::new();
-        let room_name = "some room";
-
-        server.create_new_room(None, room_name.to_owned().clone());
-        let player_id: PlayerID = {
-            let player: &mut Player = server.add_new_player("some player".to_owned(), fake_socket_addr());
-            player.player_id
-        };
-
-        server.join_room(player_id, room_name.to_owned());
-
-        server.handle_chat_message(player_id, "Conwayste is such a fun game".to_owned());
-        server.handle_chat_message(player_id, "There are not loot boxes".to_owned());
-        server.handle_chat_message(player_id, "It is free!".to_owned());
-        server.handle_chat_message(player_id, "What's not to love?".to_owned());
-
-        let current_timestamp = time::Instant::now();
-        let travel_to_the_past = current_timestamp - Duration::from_secs((MAX_AGE_CHAT_MESSAGES+1) as u64);
-        for ref mut room in server.rooms.values_mut() {
-            println!("Room: {:?}", room.name);
-            for m in room.messages.iter_mut() {
-                println!("{:?}, {:?},       {:?}", m.timestamp, travel_to_the_past, m.timestamp - travel_to_the_past);
-                m.timestamp = travel_to_the_past;
-            }
-        }
-
-        // Messages are not old enough to be expired
-        server.expire_old_messages_in_all_rooms();
-
-        for room in server.rooms.values() {
-            assert_eq!(room.messages.len(), 0);
-        }
-    }
-
-    #[test]
-    fn expire_old_messages_in_all_rooms_several_rooms_old_messages_are_wiped()
-    {
-        let mut server = ServerState::new();
-        let room_name = "some room";
-        let room_name2 = "some room 2";
-
-        server.create_new_room(None, room_name.to_owned().clone());
-        let player_id: PlayerID = {
-            let player: &mut Player = server.add_new_player("some player".to_owned(), fake_socket_addr());
-            player.player_id
-        };
-        server.create_new_room(None, room_name2.to_owned().clone());
-        let player_id2: PlayerID = {
-            let player: &mut Player = server.add_new_player("some player".to_owned(), fake_socket_addr());
-            player.player_id
-        };
-
-        server.join_room(player_id, room_name.to_owned());
-        server.join_room(player_id2, room_name.to_owned());
-
-        server.handle_chat_message(player_id, "Conwayste is such a fun game".to_owned());
-        server.handle_chat_message(player_id, "There are not loot boxes".to_owned());
-        server.handle_chat_message(player_id2, "It is free!".to_owned());
-        server.handle_chat_message(player_id2, "What's not to love?".to_owned());
-
-        let current_timestamp = time::Instant::now();
-        let travel_to_the_past = current_timestamp - Duration::from_secs((MAX_AGE_CHAT_MESSAGES+1) as u64);
-        for ref mut room in server.rooms.values_mut() {
-            println!("Room: {:?}", room.name);
-            for m in room.messages.iter_mut() {
-                println!("{:?}, {:?},       {:?}", m.timestamp, travel_to_the_past, m.timestamp - travel_to_the_past);
-                m.timestamp = travel_to_the_past;
-            }
-        }
-
-        // Messages are not old enough to be expired
-        server.expire_old_messages_in_all_rooms();
-
-        for room in server.rooms.values() {
-            assert_eq!(room.messages.len(), 0);
-        }
-    }
-
-    #[test]
-    fn handle_new_connection_good_case() {
-        let mut server = ServerState::new();
-        let player_name = "some name".to_owned();
-        let pkt = server.handle_new_connection(player_name, fake_socket_addr());
-        match pkt {
-            Packet::Response{sequence: _, request_ack: _, code} => {
-                match code {
-                    ResponseCode::LoggedIn(_,_) => {}
-                    _ => panic!("Unexpected ResponseCode: {:?}", code)
-                }
-            }
-            _ => panic!("Unexpected Packet Type: {:?}", pkt)
-        }
-    }
-
-    #[test]
-    fn handle_new_connection_player_name_taken() {
-        let mut server = ServerState::new();
-        let player_name = "some name".to_owned();
-
-        let pkt = server.handle_new_connection(player_name.clone(), fake_socket_addr());
-        match pkt {
-            Packet::Response{sequence: _, request_ack: _, code} => {
-                match code {
-                    ResponseCode::LoggedIn(_,version) => {assert_eq!(version, net::VERSION.to_owned() )}
-                    _ => panic!("Unexpected ResponseCode: {:?}", code)
-                }
-            }
-            _ => panic!("Unexpected Packet Type: {:?}", pkt)
-        }
-
-        let pkt = server.handle_new_connection(player_name, fake_socket_addr());
-        match pkt {
-            Packet::Response{sequence: _, request_ack: _, code} => {
-                match code {
-                    ResponseCode::Unauthorized(msg) => { assert_eq!(msg, Some("not a unique name".to_owned())); }
-                    _ => panic!("Unexpected ResponseCode: {:?}", code)
-                }
-            }
-            _ => panic!("Unexpected Packet Type: {:?}", pkt)
-        }
-    }
-
-    fn a_request_action_strat() -> BoxedStrategy<RequestAction> {
-        prop_oneof![
-            //Just(RequestAction::Disconnect), // not yet implemented
-            //Just(RequestAction::KeepAlive),  // same
-            Just(RequestAction::LeaveRoom),
-            Just(RequestAction::ListPlayers),
-            Just(RequestAction::ListRooms),
-            Just(RequestAction::None),
-        ].boxed()
-    }
-
-    fn a_request_action_complex_strat() -> BoxedStrategy<RequestAction> {
-        prop_oneof![
-            ("([A-Z]{1,4} [0-9]{1,2}){3}").prop_map(|a| RequestAction::ChatMessage(a)),
-            ("([A-Z]{1,4} [0-9]{1,2}){3}").prop_map(|a| RequestAction::NewRoom(a)),
-            ("([A-Z]{1,4} [0-9]{1,2}){3}").prop_map(|a| RequestAction::JoinRoom(a)),
-            ("([A-Z]{1,4} [0-9]{1,2}){3}", "[0-9].[0-9].[0-9]").prop_map(|(a, b)| RequestAction::Connect{name: a, client_version: b})
-        ].boxed()
-    }
-
-    // These tests are checking that we do not panic on each RequestAction
-    proptest! {
-        #[test]
-        fn process_request_action_simple(ref request in a_request_action_strat()) {
-            let mut server = ServerState::new();
-            server.create_new_room(None, "some room".to_owned().clone());
-            let player_id: PlayerID = {
-                let player: &mut Player = server.add_new_player("some player".to_owned(), fake_socket_addr());
-                player.player_id
-            };
-            server.process_request_action(player_id, request.to_owned());
-        }
-
-        #[test]
-        fn process_request_action_complex(ref request in a_request_action_complex_strat()) {
-            let mut server = ServerState::new();
-            server.create_new_room(None, "some room".to_owned().clone());
-            let player_id: PlayerID = {
-                let player: &mut Player = server.add_new_player("some player".to_owned(), fake_socket_addr());
-                player.player_id
-            };
-            server.process_request_action(player_id, request.to_owned());
-        }
-    }
-
-    #[test]
-    fn process_request_action_connect_while_connected() {
-        let mut server = ServerState::new();
-        server.create_new_room(None, "some room".to_owned().clone());
-        let player_id: PlayerID = {
-            let player: &mut Player = server.add_new_player("some player".to_owned(), fake_socket_addr());
-            player.player_id
-        };
-        let result = server.process_request_action(player_id, RequestAction::Connect{name: "some player".to_owned(), client_version: "0.1.0".to_owned()});
-        assert_eq!(result, ResponseCode::BadRequest(Some("already connected".to_owned())));
-    }
-
-    #[test]
-    fn process_request_action_none_is_invalid() {
-        let mut server = ServerState::new();
-        server.create_new_room(None, "some room".to_owned().clone());
-        let player_id: PlayerID = {
-            let player: &mut Player = server.add_new_player("some player".to_owned(), fake_socket_addr());
-            player.player_id
-        };
-        let result = server.process_request_action(player_id, RequestAction::None);
-        assert_eq!(result, ResponseCode::BadRequest(Some("Invalid request".to_owned())));
-    }
-
-    #[test]
-    fn prepare_response_spot_check_response_packet() {
-        let mut server = ServerState::new();
-        let player_id: PlayerID = {
-            let player: &mut Player = server.add_new_player("some player".to_owned(), fake_socket_addr());
-            player.request_ack = Some(1);
-            player.player_id
-        };
-        let pkt: Packet = server.prepare_response(player_id, RequestAction::ListRooms).unwrap();
-        match pkt {
-            Packet::Response{code, sequence, request_ack} => {
-                assert_eq!(code, ResponseCode::RoomList(vec![]));
-                assert_eq!(sequence, 1);
-                assert_eq!(request_ack, Some(2));
-            }
-            _ => panic!("Unexpected Packet type on Response path: {:?}", pkt)
-        }
-        let player: &Player = server.get_player(player_id);
-        assert_eq!(player.next_resp_seq, 2);
-    }
-
-    #[test]
-    fn validate_client_version_client_is_up_to_date() {
-        assert_eq!(validate_client_version( env!("CARGO_PKG_VERSION").to_owned()), true);
-    }
-
-    #[test]
-    fn validate_client_version_client_is_very_old() {
-      assert_eq!(validate_client_version("0.0.1".to_owned()), true);
-    }
-
-    #[test]
-    fn validate_client_version_client_is_from_the_future() {
-        assert_eq!(validate_client_version(format!("{}.{}.{}", <i32>::max_value(), <i32>::max_value(), <i32>::max_value()).to_owned()), false);
-    }
-
-    #[test]
-    fn decode_packet_update_reply_good_case() {
-        let mut server = ServerState::new();
-        let cookie: String = {
-            let player: &mut Player = server.add_new_player("some player".to_owned(), fake_socket_addr());
-            player.cookie.clone()
-        };
-
-        let update_reply_packet = Packet::UpdateReply {
-                cookie: cookie,
-                last_chat_seq: Some(0),
-                last_game_update_seq: None,
-                last_gen: None,
-        };
-
-        let result = server.decode_packet(fake_socket_addr(), update_reply_packet);
-        assert!(result.is_ok());
-        assert!(result.unwrap().is_none());
-    }
-
-    #[test]
-    fn decode_packet_update_reply_invalid_cookie() {
-        let mut server = ServerState::new();
-        {
-            let player: &mut Player = server.add_new_player("some player".to_owned(), fake_socket_addr());
-            player.cookie.clone()
-        };
-
-        let cookie = "CookieMonster".to_owned();
-
-        let update_reply_packet = Packet::UpdateReply {
-                cookie: cookie,
-                last_chat_seq: Some(0),
-                last_game_update_seq: None,
-                last_gen: None,
-        };
-
-        let result = server.decode_packet(fake_socket_addr(), update_reply_packet);
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn construct_client_updates_no_rooms() {
-        let mut server = ServerState::new();
-        let result = server.construct_client_updates();
-        assert!(result.is_ok());
-        assert!(result.unwrap().is_none());
-    }
-
-    #[test]
-    fn construct_client_updates_empty_rooms() {
-        let mut server = ServerState::new();
-        server.create_new_room(None, "some room".to_owned().clone());
-        let result = server.construct_client_updates();
-        assert!(result.is_ok());
-        assert!(result.unwrap().is_none());
-    }
-
-    #[test]
-    fn construct_client_updates_populated_room_returns_all_messages() {
-        let mut server = ServerState::new();
-        let room_name = "some_room".to_owned();
-        let player_name = "some player".to_owned();
-        let message_text = "Message".to_owned();
-
-        server.create_new_room(None, room_name.clone());
-
-        let player_id: PlayerID = {
-            let player: &mut Player = server.add_new_player(player_name.clone(), fake_socket_addr());
-            player.player_id
-        };
-        server.join_room(player_id, room_name);
-        server.handle_chat_message(player_id, message_text.clone());
-        server.handle_chat_message(player_id, message_text.clone());
-        server.handle_chat_message(player_id, message_text.clone());
-        let result = server.construct_client_updates();
-
-        assert!(result.is_ok());
-        let opt_output = result.unwrap();
-        assert!(opt_output.is_some());
-        let mut output: Vec<(SocketAddr, Packet)> = opt_output.unwrap();
-
-        // Vector should contain a single item for this test
-        assert_eq!(output.len(), 1);
-
-        let (addr, pkt) = output.pop().unwrap();
-        assert_eq!(addr, fake_socket_addr());
-
-        match pkt {
-            Packet::Update{chats, game_updates, universe_update} => {
-                assert_eq!(game_updates, None);
-                assert_eq!(universe_update, UniUpdateType::NoChange);
-                assert!(chats.is_some());
-
-                // All client chat sequence numbers start counting at 1
-                let mut i=1;
-
-                for msg in chats.unwrap() {
-                    assert_eq!(msg.player_name, player_name);
-                    assert_eq!(msg.chat_seq, Some(i));
-                    assert_eq!(msg.message, message_text);
-                    i+=1;
-                }
-            }
-            _ => panic!("Unexpected packet in client update construction!")
-        }
-    }
-
-    #[test]
-    fn construct_client_updates_populated_room_returns_updates_after_client_acked() {
-        let mut server = ServerState::new();
-        let room_name = "some_room".to_owned();
-        let player_name = "some player".to_owned();
-        let message_text = "Message".to_owned();
-
-        server.create_new_room(None, room_name.clone());
-
-        let player_id: PlayerID = {
-            let player: &mut Player = server.add_new_player(player_name.clone(), fake_socket_addr());
-            player.player_id
-        };
-        server.join_room(player_id, room_name);
-        server.handle_chat_message(player_id, message_text.clone());
-        server.handle_chat_message(player_id, message_text.clone());
-        server.handle_chat_message(player_id, message_text.clone());
-
-        // Assume that the client has acknowledged two chats
-        {
-            let player: &mut Player = server.get_player_mut(player_id);
-            player.update_chat_seq_num(Some(2));
-        }
-
-        // We should then only return the last chat
-        let result = server.construct_client_updates();
-
-        assert!(result.is_ok());
-        let opt_output = result.unwrap();
-        assert!(opt_output.is_some());
-        let mut output: Vec<(SocketAddr, Packet)> = opt_output.unwrap();
-
-        // Vector should contain a single item for this test
-        assert_eq!(output.len(), 1);
-
-        let (addr, pkt) = output.pop().unwrap();
-        assert_eq!(addr, fake_socket_addr());
-
-        match pkt {
-            Packet::Update{chats, game_updates, universe_update} => {
-                assert_eq!(game_updates, None);
-                assert_eq!(universe_update, UniUpdateType::NoChange);
-                assert!(chats.is_some());
-
-                let mut messages = chats.unwrap();
-                assert_eq!(messages.len(), 1);
-                let msg = messages.pop().unwrap();
-
-                assert_eq!(msg.player_name, player_name);
-                assert_eq!(msg.chat_seq, Some(3));
-                assert_eq!(msg.message, message_text);
-            }
-            _ => panic!("Unexpected packet in client update construction!")
-        }
-    }
-
-    #[test]
-    fn broadcast_message_to_two_players_in_room() {
-        let mut server = ServerState::new();
-        let room_name = "some_room".to_owned();
-        let player_name = "some player".to_owned();
-
-        server.create_new_room(None, room_name.clone());
-
-        let player_id: PlayerID = {
-            let player: &mut Player = server.add_new_player(player_name.clone(), fake_socket_addr());
-            player.player_id
-        };
-        let player_id2: PlayerID = {
-            let player: &mut Player = server.add_new_player(player_name.clone(), fake_socket_addr());
-            player.player_id
-        };
-
-        server.join_room(player_id, room_name.clone());
-        {
-            let room: &mut Room = server.get_room_mut(player_id).unwrap();
-            room.broadcast("Silver birch against a Swedish sky".to_owned());
-        }
-        server.join_room(player_id2, room_name.clone());
-        let room: &Room = server.get_room(player_id).unwrap();
-
-        let player = (*server.get_player(player_id)).clone();
-        let msgs = server.collect_unacknowledged_messages(room, &player).unwrap();
-        assert_eq!(msgs.len(), 1);
-        assert_eq!(msgs[0].message, "Silver birch against a Swedish sky".to_owned());
-
-        let player = (*server.get_player(player_id2)).clone();
-        let msgs = server.collect_unacknowledged_messages(room, &player).unwrap();
-        assert_eq!(msgs.len(), 1);
-        assert_eq!(msgs[0].message, "Silver birch against a Swedish sky".to_owned());
-    }
-
-    #[test]
-    fn broadcast_message_to_an_empty_room() {
-        let mut server = ServerState::new();
-        let room_name = "some_room".to_owned();
-
-        server.create_new_room(None, room_name.clone());
-        let room_id: &RoomID = server.room_map.get(&room_name.clone()).unwrap();
-
-        {
-            let room: &mut Room = server.rooms.get_mut(&room_id).unwrap();
-            room.broadcast("Silver birch against a Swedish sky".to_owned());
-        }
-        let room: &Room = server.rooms.get(&room_id).unwrap();
-        assert_eq!(room.latest_seq_num, 1);
-        assert_eq!(room.messages.len(), 1);
-        let msgs: &ServerChatMessage = room.messages.get(0).unwrap();
-        assert_eq!(msgs.player_name, "Server".to_owned());
-        assert_eq!(msgs.seq_num, 1);
-        assert_eq!(msgs.player_id, PlayerID(0xFFFFFFFFFFFFFFFF));
-    }
-
-    #[test]
-    #[should_panic]
-    fn disconnect_get_player_by_id_fails() {
-        let mut server = ServerState::new();
-        let player_name = "some player".to_owned();
-
-        let player_id: PlayerID = {
-            let player: &mut Player = server.add_new_player(player_name.clone(), fake_socket_addr());
-            player.player_id
-        };
-
-        server.handle_disconnect(player_id);
-        server.get_player(player_id);
-    }
-
-    #[test]
-    fn disconnect_get_player_by_cookie_fails() {
-        let mut server = ServerState::new();
-        let player_name = "some player".to_owned();
-
-        let (player_id, cookie) = {
-            let player: &mut Player = server.add_new_player(player_name.clone(), fake_socket_addr());
-            (player.player_id, player.cookie.clone())
-        };
-
-        server.handle_disconnect(player_id);
-        assert_eq!(server.get_player_id_by_cookie(cookie.as_str()), None);
-    }
-
-    #[test]
-    fn disconnect_while_in_room_removes_all_traces_of_player() {
-        let mut server = ServerState::new();
-        let room_name = "some_room".to_owned();
-        let player_name = "some player".to_owned();
-
-        let player_id: PlayerID = {
-            let player: &mut Player = server.add_new_player(player_name.clone(), fake_socket_addr());
-            player.player_id
-        };
-
-        server.create_new_room(None, room_name.clone());
-        server.join_room(player_id, room_name);
-        let room_id = {
-            let room: &Room = server.get_room(player_id).unwrap();
-            assert_eq!(room.player_ids.contains(&player_id), true);
-            room.room_id
-        };
-        server.handle_disconnect(player_id);
-        // Cannot go through player_id because the player has been removed
-        let room: &Room = server.rooms.get(&room_id).unwrap();
-        assert_eq!(room.player_ids.contains(&player_id), false);
-    }
-
-    #[test]
-    fn test_is_previously_processed_packet() {
-        let mut server = ServerState::new();
-        let player_name = "some player".to_owned();
-
-        let player_id: PlayerID = {
-            let player: &mut Player = server.add_new_player(player_name.clone(), fake_socket_addr());
-            player.request_ack = Some(4);
-            player.player_id
-        };
-
-        let player_id2: PlayerID = {
-            let player: &mut Player = server.add_new_player(player_name.clone(), fake_socket_addr());
-            player.request_ack = None;
-            player.player_id
-        };
-
-        assert_eq!(server.is_previously_processed_packet(player_id2, 0), false);
-
-        assert_eq!(server.is_previously_processed_packet(player_id, 0), true);
-        assert_eq!(server.is_previously_processed_packet(player_id, 4), true);
-        assert_eq!(server.is_previously_processed_packet(player_id, 5), false);
-    }
-
-    #[test]
-    fn test_clear_transmission_queue_on_ack() {
-        let mut server = ServerState::new();
-        let player_name = "some player".to_owned();
-
-        let player_id: PlayerID = {
-            let player: &mut Player = server.add_new_player(player_name.clone(), fake_socket_addr());
-            player.request_ack = Some(4);
-            player.player_id
-        };
-
-        for i in 0..5 {
-            let pkt = Packet::Response {
-                sequence: i,
-                request_ack: None,
-                code: ResponseCode::OK
-            };
-
-            let nm: &mut NetworkManager = server.network_map.get_mut(&player_id).unwrap();
-            nm.tx_packets.buffer_item(pkt.clone());
-        }
-
-        server.clear_transmission_queue_on_ack(player_id, None);
-        assert_eq!(server.network_map.get(&player_id).unwrap().tx_packets.len(), 5);
-        server.clear_transmission_queue_on_ack(player_id, Some(0));
-        assert_eq!(server.network_map.get(&player_id).unwrap().tx_packets.len(), 5);
-        server.clear_transmission_queue_on_ack(player_id, Some(1));
-        assert_eq!(server.network_map.get(&player_id).unwrap().tx_packets.len(), 4);
-        server.clear_transmission_queue_on_ack(player_id, Some(5));
-        assert_eq!(server.network_map.get(&player_id).unwrap().tx_packets.len(), 0);
-    }
-
-    #[test]
-    fn test_resend_expired_tx_packets_empty_server() {
-        let mut server = ServerState::new();
-
-        let (udp_tx, _) = mpsc::unbounded();
-        #[cfg(not(should_panic))]
-        server.resend_expired_tx_packets(&udp_tx);
-    }
-
-    #[test]
-    fn test_resend_expired_tx_packets() {
-        let mut server = ServerState::new();
-        let player_name = "some player".to_owned();
-
-        let player_id: PlayerID = {
-            let player: &mut Player = server.add_new_player(player_name.clone(), fake_socket_addr());
-            player.request_ack = Some(5);
-            player.player_id
-        };
-
-
-        for i in 0..5 {
-            let pkt = Packet::Response {
-                sequence: i,
-                request_ack: None,
-                code: ResponseCode::OK
-            };
-
-            let nm: &mut NetworkManager = server.network_map.get_mut(&player_id).unwrap();
-            nm.tx_packets.buffer_item(pkt.clone());
-
-            if i < 3 {
-                let attempt: &mut net::NetAttempt = nm.tx_packets.attempts.back_mut().unwrap();
-                attempt.time = Instant::now() - Duration::from_secs(i+1);
-            }
-        }
-
-        let (udp_tx, _) = mpsc::unbounded();
-        server.resend_expired_tx_packets(&udp_tx);
-
-        for i in 0..5 {
-            let nm: &mut NetworkManager = server.network_map.get_mut(&player_id).unwrap();
-            let packet_retries: &net::NetAttempt = nm.tx_packets.attempts.get(i).unwrap();
-
-            if i >= 3 {
-                assert_eq!(packet_retries.retries, 0);
-            } else {
-                assert_eq!(packet_retries.retries, 1);
-            }
-        }
-    }
-
-    #[test]
-    fn test_process_queued_rx_packets_first_non_connect_player_packet() {
-        let mut server = ServerState::new();
-        let player_name = "some player".to_owned();
-
-        let (player_id, player_cookie): (PlayerID, String) = {
-            let player: &mut Player = server.add_new_player(player_name.clone(), fake_socket_addr());
-            player.request_ack = Some(1);   // Player connected and we've confirmed the first transaction
-            (player.player_id, player.cookie.clone())
-        };
-
-        {
-            let pkt = Packet::Request {
-                cookie: Some(player_cookie),
-                sequence: 2,
-                response_ack: None,
-                action: RequestAction::ListPlayers
-            };
-
-            let nm: &mut NetworkManager = server.network_map.get_mut(&player_id).unwrap();
-            nm.rx_packets.buffer_item(pkt.clone());
-
-            assert_eq!(nm.tx_packets.len(), 0);
-        }
-
-        server.process_queued_rx_packets(player_id);
-
-
-        {
-            let nm: &mut NetworkManager = server.network_map.get_mut(&player_id).unwrap();
-            assert_eq!(nm.tx_packets.len(), 1);
-        }
-
-    }
-
-    #[test]
-    fn test_process_queued_rx_packets_contiguous() {
-        let mut server = ServerState::new();
-        let player_name = "some player".to_owned();
-
-        let (player_id, player_cookie): (PlayerID, String) = {
-            let player: &mut Player = server.add_new_player(player_name.clone(), fake_socket_addr());
-            player.request_ack = Some(1);   // Player connected and we've confirmed the first transaction
-            (player.player_id, player.cookie.clone())
-        };
-
-        for i in 2..10 {
-            let pkt = Packet::Request {
-                cookie: Some(player_cookie.clone()),
-                sequence: i,
-                response_ack: None,
-                action: RequestAction::ListPlayers
-            };
-
-            let nm: &mut NetworkManager = server.network_map.get_mut(&player_id).unwrap();
-            nm.rx_packets.buffer_item(pkt.clone());
-
-            assert_eq!(nm.tx_packets.len(), 0);
-        }
-
-        server.process_queued_rx_packets(player_id);
-
-        {
-            let nm: &mut NetworkManager = server.network_map.get_mut(&player_id).unwrap();
-            assert_eq!(nm.tx_packets.len(), 8);
-        }
-
-    }
-
-    #[test]
-    fn test_process_queued_rx_packets_swiss_cheese_queue() {
-        let mut server = ServerState::new();
-        let player_name = "some player".to_owned();
-
-        let (player_id, player_cookie): (PlayerID, String) = {
-            let player: &mut Player = server.add_new_player(player_name.clone(), fake_socket_addr());
-            player.request_ack = Some(1);   // Player connected and we've confirmed the first transaction
-            (player.player_id, player.cookie.clone())
-        };
-
-        for i in [2, 3, 4, 6, 8, 9, 10].iter() {
-            let pkt = Packet::Request {
-                cookie: Some(player_cookie.clone()),
-                sequence: *i,
-                response_ack: None,
-                action: RequestAction::ListPlayers
-            };
-
-            let nm: &mut NetworkManager = server.network_map.get_mut(&player_id).unwrap();
-            nm.rx_packets.buffer_item(pkt.clone());
-
-            assert_eq!(nm.tx_packets.len(), 0);
-        }
-
-        server.process_queued_rx_packets(player_id);
-
-        {
-            let nm: &mut NetworkManager = server.network_map.get_mut(&player_id).unwrap();
-            assert_eq!(nm.tx_packets.len(), 3); // only 2, 3, and 4 are processed
         }
     }
 }
